requirements.txt
Flask~=3.1.0
mysql-connector-python~=9.3.0
gunicorn~=23.0.0